test_task/backend/backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg==3.0.11
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
